Made gui_to_map() to use floats internally.
See patch #6447

[[originally from svn r30297]]

+++ client/mapview_common.h
@@ -249,8 +249,8 @@ void unqueue_mapview_updates(bool write_to_screen);
 void map_to_gui_vector(const struct tileset *t, float zoom,
 		       float *gui_dx, float *gui_dy, int map_dx, int map_dy);
 bool tile_to_canvas_pos(float *canvas_x, float *canvas_y, struct tile *ptile);
-struct tile *canvas_pos_to_tile(int canvas_x, int canvas_y);
-struct tile *canvas_pos_to_nearest_tile(int canvas_x, int canvas_y);
+struct tile *canvas_pos_to_tile(float canvas_x, float canvas_y);
+struct tile *canvas_pos_to_nearest_tile(float canvas_x, float canvas_y);
 
 void get_mapview_scroll_window(float *xmin, float *ymin,
                                float *xmax, float *ymax,
